Update so output files include ensemble ids
Both ensemble and sensitivity analysis now save files with
ensemble id in the file name. This prevents previous analyses
from being automatically overwritten. In particular, when
iteratively updating an analysis it is useful to be able to see
how model output is changing.

This didn't seem to affect much of the code, but I grepped for and
fixed the few places where these files need to be accessed by name.

Currently, the ensemble id needed to name/access files gets passed
around in the settings list. This isn't ideal, since it gets overwritten
any time a new analysis is performed. Thus PEcAn will always look
for the most recent files, which is OK, but better might be if we
kept a record of all the ensemble IDs generated within the workflow.
I'll work on that in the future, but for now this is a good step.