Allow parameters set as constant in pecan.xml to stay put throughout analysis
Maybe I was missing something about how this is supposed to work, but I couldn't
get what I thought was the desired behavior to work for me. I want to specify constant
parameters for one or more PFTs in pecan.xml, and then have these parameters remain
at the values I set for any and all analyses. Seems to be working for me now.