Fix processing of defaults names in ed config.xml
This introduces a failsafe that should modify the names of `defaults`
(`settings$constants`) based on the sub-list item `defaults$name`. This
only runs if the `defaults` have no names (`is.null(names(defaults))`)
or if any of the names are the uninformative `pft` tag. Through the new
syntax, this will ONLY change names that are not `NULL` -- places where
they are remain unaltered.

I also made the test script more robust, with a `settings` list that
contains a constant,