Make sure `include_pft` is correct in write.config.xml.ED2
Values for the config.xml are read in from history files:

`vals <- as.list(edhistory[edhistory$num == pft.number, ])`

But depending on the history file, `include_pft` may be set to 0. However, I'm assuming that if you included the PFT in your `pecan.xml` then you definitely do want `include_pft = 1`. Thus it seems safe to set it to 1 at the end of conversions.